feat: rearrange PredictedSchedule.group/3 to avoid multiple loops
Previously, we'd loop over the schedules and predictions twice: once to build
the list of keys, and another time to do the merge. We don't need to do that!

We can iterate through each list once, building the map as we go. Schedules
is generally larger, so we do that one first to give predictions a better
chance of matching an existing map key.